requirements.txt
PyMuPDF
python-docx
pytesseract
pillow
openai
tqdm
requests
beautifulsoup4
python-dotenv
schedule
